+++ src/main/java/ar/edu/utn/dds/k3003/model/SuscripcionCantViandasRequest.java
@@ -7,5 +7,5 @@ public class SuscripcionCantViandasRequest {
     @Getter
     private int cantidadViandas;
     @Getter
-    private Long heladera_id;
+    private int heladera_id;
 }
